refactor CriteriaInput component to own file
